Backend/resource_operation/requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
azure-cosmosdb-table
msrestazure